requirements.txt
streamlit
pymongo
pydantic
pydantic-settings
python-dotenv
google-genai